requirements.txt
Django==2.0.1
Pillow==4.2.1
PyYAML==3.12
chardet==2.2.1
colorama==0.2.5
diff-match-patch==20121119
django-appconf==1.0.2
django-extensions==1.9.1
django-filter==1.1.0
django-forms-bootstrap==3.1.0
django-haystack==2.8.1
django-js-asset==1.0.0
django-mptt==0.9.0
et-xmlfile==1.0.1
html5lib==0.999
jdcal==1.3
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.0
psycopg2==2.7.5
pycurl==7.19.3
pygobject==3.12.0
python-apt===0.9.3.5ubuntu2
pytz==2017.3
requests==2.2.1
six==1.5.2
tablib==0.12.1
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.7.1
wheel==0.24.0
xlrd==1.1.0
xlwt==1.3.0
